refactor: Drastically simplify package __init__.py (Option C)
MAJOR SIMPLIFICATION:
- Removed ALL imports from __init__.py
- Removed offline mode setup from package initialization
- Removed conditional imports
- Now just: version assignment + __all__ declaration (32 lines vs 97)

Moved offline mode setup:
- Created standalone offline_mode.py utility module
- Updated onnx_ranker.py to call setup_offline_mode() before ML imports
- Only runs when actually needed, not during package import

Benefits:
- No circular import possibility
- Package initialization is trivial and fast
- Reduces pytest collection complexity
- Offline mode still works where needed

This eliminates the root cause of package recognition issues by
making __init__.py as simple as possible - just version and exports.
